debug_rel branch: CMakeLists.txt: automate naming of libraries Sacado needs. checkpoint.cpp: fix warnings.

+++ test_more/general/checkpoint.cpp
@@ -17,14 +17,14 @@ namespace {
 
 	bool f_algo(const ADVector& x, ADVector& y)
 	{	size_t m = y.size();
-		assert( x.size() == m + 1);
+		assert( size_t(x.size()) == m + 1);
 		for(size_t i = 0; i < m; i++)
 			y[i] = x[i] * x[i+1];
 		return true;
 	}
 	bool g_algo(const ADVector& y, ADVector& z)
 	{	size_t m = z.size();
-		assert( y.size() + 1 == m );
+		assert( size_t(y.size()) + 1 == m );
 		z[0] = 0.0;
 		for(size_t i = 1; i < m; i++)
 		{	z[0] += y[i-1];
